Тест создания канала через главную кнопку
