+++ pages/base.py
@@ -1,0 +1,17 @@
+from urlparse import  urljoin
+
+class Page(object):
+    BASE_URL = 'https://ok.ru/'
+    PATH = ''
+
+    def __init__(self, driver):
+        self.driver = driver
+
+    def open(self):
+        url = urljoin(self.BASE_URL, self.PATH)
+        self.driver.get(url)
+        self.driver.maximize_window()
+
+class Component(object):
+    def __init__(self, driver):
+        self.driver = driver
